geo_back/pyproject.toml
[tool.poetry]
name = "geo-back"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.1"
djangorestframework = "3.13.1"
psycopg2-binary = "^2.9.9"
environs = "^10.3.0"
networkx = "^3.2.1"
osmnx = "^1.9.1"
scipy = "^1.12.0"
djangorestframework-gis = "^1.0"
sqlalchemy = "^2.0.27"
geoalchemy2 = "^0.14.4"
django-debug-toolbar = "^4.3.0"
django-cors-headers = "^4.3.1"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.1"
pytest-factoryboy = "^2.7.0"
pytest-django = "^4.8.0"
pytest-cov = "^5.0.0"
black = "^24.3.0"
pre-commit = "^3.7.0"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
